requirements.txt
beautifulsoup4            4.7.1     
bs4                       0.0.1     
certifi                   2018.11.29
chardet                   3.0.4     
coreapi                   2.3.3     
coreschema                0.0.4     
coverage                  4.5.3     
Django                    2.1.5     
django-classy-tags        0.8.0     
django-cms                3.6.0     
django-filter             2.1.0     
django-formtools          2.1       
django-oauth-toolkit      1.2.0     
django-sekizai            0.10.0    
django-treebeard          4.3       
djangocms-admin-style     1.3.0     
djangorestframework       3.9.1     
djangorestframework-oauth 1.1.0     
get                       2019.3.22 
httpie                    1.0.2     
idna                      2.8       
itypes                    1.1.0     
Jinja2                    2.10      
Markdown                  3.0.1     
MarkupSafe                1.1.1     
oauthlib                  3.0.1     
pip                       19.0.3    
post                      2019.3.22 
psycopg2                  2.7.7     
public                    2019.3.22 
Pygments                  2.3.1     
pymongo                   3.7.2     
pytz                      2018.9    
query-string              2019.3.22 
razorpay                  1.1.1     
request                   2019.3.22 
requests                  2.21.0    
selenium                  3.141.0   
setuptools                40.7.1    
soupsieve                 1.8       
uritemplate               3.0.0     
urllib3                   1.24.1    
wheel                     0.32.3    
